feature: Added OAuth Security + File Uploads Tied to User (#43)
* updated deps

* updated pyproject.toml with latest versions

* following fastapi security to add oauth. wip

* added table to database for userinfo and added token endpoint. todo: implement ability to create new users

* created method to add new user

* removed 'disabled' field and require username to be unique

* using bcrypt to check password

* access tokens are being generated

* fix: time delta param now used if provided

* removed unused files

* added user_id as foreign key in filemetadatarecord table

* updated test for read_root to pass auth

* added auth to file endpoints and updated tests. TODO: update query logic to only return files user owns

* FileUploads now contain the user_id of the user who uploaded them

* DbClient get_metadata operation now requires user_id

* db queries are guarded by user_id of current user

* added dummy auth secret key to use in tests

* added code coverage reporting

* added new test file to test security functionality + updated coverage config

* storing coverage report as artifact

* Revert "storing coverage report as artifact"

This reverts commit ec5f89e7966939108063b919e1685393440d331e.

+++ tests/test_delete_file.py
@@ -8,13 +8,18 @@ from smolvault.models import FileUploadDTO
 
 @pytest.mark.asyncio()
 @pytest.mark.usefixtures("_test_bucket")
-async def test_delete_file(client: AsyncClient, camera_img: bytes) -> None:
+async def test_delete_file(client: AsyncClient, camera_img: bytes, access_token: str) -> None:
     # first upload the file
     filename = f"{uuid4().hex[:6]}-camera.png"
-    expected_obj = FileUploadDTO(name=filename, size=len(camera_img), content=camera_img, tags="camera,photo")
+    expected_obj = FileUploadDTO(
+        name=filename, size=len(camera_img), content=camera_img, tags="camera,photo", user_id=1
+    )
     expected = expected_obj.model_dump(exclude={"content", "upload_timestamp", "tags"})
     response = await client.post(
-        "/file/upload", files={"file": (filename, camera_img, "image/png")}, data={"tags": "camera,photo"}
+        "/file/upload",
+        files={"file": (filename, camera_img, "image/png")},
+        data={"tags": "camera,photo"},
+        headers={"Authorization": f"Bearer {access_token}"},
     )
     actual: dict[str, Any] = response.json()
     actual.pop("upload_timestamp")
@@ -22,7 +27,7 @@ async def test_delete_file(client: AsyncClient, camera_img: bytes) -> None:
     assert actual == expected
 
     # now delete the file
-    response = await client.delete(f"/file/{filename}")
+    response = await client.delete(f"/file/{filename}", headers={"Authorization": f"Bearer {access_token}"})
     actual = response.json()
     assert response.status_code == 200
     assert actual["message"] == "File deleted successfully"
